requirements.txt
ipykernel==6.16.2
ipython==8.6.0
jupyter_client==7.4.4
jupyter_core==4.11.2
pandas==1.5.1
python-dotenv==1.0.1
requests==2.28.1
